requirements.txt
# Core dependencies
python-dotenv>=1.0.0
numpy>=1.24.0
pyee==11.0.1
pydub==0.25.1
sounddevice==0.4.6
python-vlc==3.0.21203
pynput==1.7.6
requests==2.31.0

# Speech recognition
SpeechRecognition==3.10.1
openai>=1.0.0
whisper==1.1.10

# Audio processing
librosa==0.10.1
soundfile==0.12.1

# Serial communication
pyserial==3.5
pyserial-asyncio==0.6

# Holocron RAG system
supabase==2.3.5
pgvector==0.2.3
tiktoken==0.6.0
beautifulsoup4==4.12.3
soupsieve==2.5
aiohttp>=3.8.0
pinecone-client>=2.2.4
sentence-transformers==2.6.1
torch==2.2.1

# Development tools
black==23.3.0
pylint==2.17.4
pytest==7.3.1
pytest-asyncio==0.21.1

# URL Collection System
tqdm>=4.65.0
pandas>=2.0.0
lxml==5.1.0

# CLI-related dependencies
click==8.1.7
rich==13.7.1

# Additional dependencies
matplotlib==3.8.3
pedalboard==0.9.16
colorama==0.4.6
pygame==2.5.2
psycopg2-binary==2.9.9
scipy==1.12.0

# New dependencies
asyncio>=3.4.3
typing-extensions==4.9.0
python-dateutil==2.8.2
backoff>=2.2.1 